+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### 0.4.0
 - Implementing WebSocket
-- Change the header `fa fa-skyatlas` font size
+- Change the header `fa fa-skyatlas` font size.
+- Change in the local storage names.
 
 ### 0.3.0
 - Add weather setting.